+++ cloud_compute_utils/scripts/install.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 set -x
 
-sudo apt update 
-sudo apt install python3.7-minimal python3-pip -y
+
+#sudo apt update 
+#sudo apt install python3.7-minimal python3-pip -y
 
 #cd wheels/
-
-python3.7 -m pip install *whl
+source activate pytorch_p38
+pip install *whl