+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.3.0
+- improve logging for running commands
+- make command-running blocking
+
 ## 0.2.5
 - prevent the ability to terminate all instances
 - misc bug improvements